Working on a different verbal morphology.
This is more in line with the previous system, and will hopefully
require less drastic changes to make it be historically accurate.

This particular commit also replaced the custom glossaries with the
`leipzig` package for linguistic glosses, which necessitated changing
the `\Q` command for conlang examples to a more sensible `\conlang`
command.